backend/requirements.txt
Django==3.2.16
django-filter==23.1
djangorestframework==3.12.4
djoser==2.1.0
pillow==10.3.0
python-dotenv==1.0.1
psycopg2-binary==2.9.3
